Update processDataStreamRequest
- Create new RequestDataStreamMessage class
- move timeInterval calculation into message class
- move MessageUtils from WenuLinkUtils.kt to MessageUtils.kt
- Rename SetMEssageIntervalCommandLong to TelemetryMessages
