+++ app/src/main/java/org/WenuLink/mavlink/controllers/NavigationController.kt
@@ -25,7 +25,6 @@ import com.MAVLink.enums.MAV_RESULT
 import com.MAVLink.enums.MAV_SEVERITY
 import io.getstream.log.taggedLogger
 import kotlin.math.roundToInt
-import org.WenuLink.adapters.MessageUtils
 import org.WenuLink.adapters.RequestMissionAction
 import org.WenuLink.adapters.RequestStartMission
 import org.WenuLink.adapters.WenuLinkCommand
@@ -33,6 +32,7 @@ import org.WenuLink.adapters.WenuLinkHandler
 import org.WenuLink.adapters.mission.MissionNode
 import org.WenuLink.adapters.mission.RepositionAction
 import org.WenuLink.mavlink.MAVLinkClient
+import org.WenuLink.mavlink.messages.MessageUtils
 import org.WenuLink.mavlink.messages.MissionStartCommandLong
 import org.WenuLink.mavlink.messages.NavLandMissionItem
 import org.WenuLink.mavlink.messages.NavTakeoffMissionItem
